fix compilation for abs test suite

+++ includes/MN_math.h
@@ -1,54 +1,45 @@
 #ifndef MN_MATH_H
 #define MN_MATH_H
-
 #include "MN_dtype.h"
-#include <stdint.h>
-
+#include <stddef.h>
 #ifdef __cplusplus
 extern "C" {
 #endif
-
 /**
  * MN ABSOLUTE OPERATOR USING C & NEON
  * @param dst   Destination buffer (must not overlap with src).
  * @param src   Source buffer.
  * @param count Number of elements.
  */
-
 /**
  * @brief Computes absolute value of float array using C.
  */
-mn_result_t mn_abs_float_c(mn_float32_t *dst, mn_float32_t *src, uint32 count);
-mn_result_t mn_abs_vec2f_c(mn_vec2f_t *dst, mn_vec2f_t *src, uint32 count);
-mn_result_t mn_abs_vec3f_c(mn_vec3f_t *dst, mn_vec3f_t *src, uint32 count);
-mn_result_t mn_abs_vec4f_c(mn_vec4f_t *dst, mn_vec4f_t *src, uint32 count);
-
+mn_result_t mn_abs_float_c(mn_float32_t *dst, mn_float32_t *src, size_t count);
+mn_result_t mn_abs_vec2f_c(mn_vec2f_t *dst, mn_vec2f_t *src, size_t count);
+mn_result_t mn_abs_vec3f_c(mn_vec3f_t *dst, mn_vec3f_t *src, size_t count);
+mn_result_t mn_abs_vec4f_c(mn_vec4f_t *dst, mn_vec4f_t *src, size_t count);
 /**
  * @brief Computes absolute value of float array using NEON.
  */
-mn_result_t mn_abs_float_neon(mn_float32_t *dst, mn_float32_t *src, uint32 count);
-mn_result_t mn_abs_vec2f_neon(mn_vec2f_t *dst, mn_vec2f_t *src, uint32 count);
-mn_result_t mn_abs_vec3f_neon(mn_vec3f_t *dst, mn_vec3f_t *src, uint32 count);
-mn_result_t mn_abs_vec4f_neon(mn_vec4f_t *dst, mn_vec4f_t *src, uint32 count);
-
+mn_result_t mn_abs_float_neon(mn_float32_t *dst, mn_float32_t *src, size_t count);
+mn_result_t mn_abs_vec2f_neon(mn_vec2f_t *dst, mn_vec2f_t *src, size_t count);
+mn_result_t mn_abs_vec3f_neon(mn_vec3f_t *dst, mn_vec3f_t *src, size_t count);
+mn_result_t mn_abs_vec4f_neon(mn_vec4f_t *dst, mn_vec4f_t *src, size_t count);
 /**
  * @brief Computes absolute value of int array using C.
  */
-mn_result_t mn_abs_int32_c(mn_int32_t *dst, mn_int32_t *src, uint32 count);
-mn_result_t mn_abs_vec2i_c(mn_vec2i_t *dst, mn_vec2i_t *src, uint32 count);
-mn_result_t mn_abs_vec3i_c(mn_vec3i_t *dst, mn_vec3i_t *src, uint32 count);
-mn_result_t mn_abs_vec4i_c(mn_vec4i_t *dst, mn_vec4i_t *src, uint32 count);
-
+mn_result_t mn_abs_int32_c(mn_int32_t *dst, mn_int32_t *src, size_t count);
+mn_result_t mn_abs_vec2i_c(mn_vec2i_t *dst, mn_vec2i_t *src, size_t count);
+mn_result_t mn_abs_vec3i_c(mn_vec3i_t *dst, mn_vec3i_t *src, size_t count);
+mn_result_t mn_abs_vec4i_c(mn_vec4i_t *dst, mn_vec4i_t *src, size_t count);
 /**
  * @brief Computes absolute value of int array using NEON.
  */
-mn_result_t mn_abs_int32_neon(mn_int32_t *dst, mn_int32_t *src, uint32 count);
-mn_result_t mn_abs_vec2i_neon(mn_vec2i_t *dst, mn_vec2i_t *src, uint32 count);
-mn_result_t mn_abs_vec3i_neon(mn_vec3i_t *dst, mn_vec3i_t *src, uint32 count);
-mn_result_t mn_abs_vec4i_neon(mn_vec4i_t *dst, mn_vec4i_t *src, uint32 count);
-
+mn_result_t mn_abs_int32_neon(mn_int32_t *dst, mn_int32_t *src, size_t count);
+mn_result_t mn_abs_vec2i_neon(mn_vec2i_t *dst, mn_vec2i_t *src, size_t count);
+mn_result_t mn_abs_vec3i_neon(mn_vec3i_t *dst, mn_vec3i_t *src, size_t count);
+mn_result_t mn_abs_vec4i_neon(mn_vec4i_t *dst, mn_vec4i_t *src, size_t count);
 #ifdef __cplusplus
 }
 #endif
-
 #endif // MN_MATH_H
